Port InputConsumerImpl to SurfaceFlinger input.
In the old system the InputConsumerImpl generated a handle
which was inserted after the appropriate window when building the list.
In the new system the InputConsumerImpl will need to manage a Surface
to assosciate this with. We hide all of them at the beginning of each Input pass
and show them again with a relative layer to the window they would have been
inserted after in the old logic. This resetting emulates the effect
of rebuilding the entire list every time.

Bug: 80101428
Bug: 113136004
Bug: 111440400
Change-Id: I387e16b82a76648d39f8cfc10cd5d30d4690275b

Former-commit-id: a80ad04b39030098a528982b5f4729495ea664e7