Restrict CAPTURE_VIDEO permissions to signature
We want to restrict screen reading permissions
to signature only so only the platform and
apps signed with the platform key can access
the device screen.

Add @removed and @hide to the permission so that it's
not in the docs any more without breaking any existing
apps referring to the permissions.

Bug: 111131054
Test: atest PermissionsHostTest
Test: make & verified generated file
out/target/common/obj/PACKAGING/hiddenapi-light-greylist.txt
contains CAPTURE_VIDEO_OUTPUT and
CAPTURE_SECURE_VIDEO_OUTPUT.

Change-Id: I451fec868e5e079dc7204fb097191a565627f939

Former-commit-id: a69e3678085520f6fccbfdb2f5602e3206ed0fe6