Whitelist com.android.proxyhandler for power
There are some scenarios under which com.android.proxyhandler is
considered by the framework as never being launched (e.g. if a PAC proxy
is added after a long wait time after an upgrade), which makes all of
its network traffic to be blackholed, due to it being subjected to the
fw_standby firewall chain. Given that all of the outgoing packets from
this app are being dropped, whenever Chrome WebView (or most other apps)
uses a PAC proxy for its networking, it is completely unable to initiate
outgoing connections.

This change whitelists com.android.proxyhandler so that this does not
happen.

Bug: 110762695
Test: dumpsys usagestats' | grep proxy
      ...
      package=com.android.proxyhandler u=0 bucket=5 reason=d ... idle=n

Change-Id: I9e4debc876cbdd2f6ba35928faff8c0beca77ae1

Former-commit-id: d2f1ca8c27871b2b2d2cf37da138a2d3723375c8