Add maxTargetSdk to @UnsupportedAppUsage annotation.
This will be used by the runtime to limit access to APIs based on their
targetSdkVersion in the manifest.

When adding new maxTargetSdk values to non-SDK interfaces, we use the
SDK version of the latest major (i.e. new letter) release of Android
that has been made public. e.g. while Q is in development, we add new
maxTargetSdk values of 28 (Build.VERSION_CODES.P), even if a P MR1 has
already been released.

We do this because:
- It reduces the number of distinct SDK levels that have to be
  considered and enforced in the runtime.
- Using past releases means that there is always a well established
  value to use.

See go/hidden-api-annotations for more context.

Test: m
Bug: 110868826
Change-Id: Idbe78510acf538ce941a9a61a64fcc0bdc4de38e

Former-commit-id: d991a408929a88c5fa37b069dd1a463657a01b6c