Make system Dialog Activities theme friendly and support dark theme
Change themes of the Dialog based activities to the newly created
internal theme Theme.Dialog.Confirmation in order to add abbility
to override theme in different configurations (television in this
case).

Overriding of Theme.DeviceDefault.Light.Dialog.Alert for TV
(values-television) would work but it will break the existing behavior
of the Light theme on TV. It is changing semantics of public API
Theme.DeviceDefault.Light. It will potentially break app that is using
light theme and override foreground/text color based on assumption it's
using light theme.

Change warning text color in HarmfulAppWarningActivity to
attr/textColorPrimary in order to support theme change (which is still
Dark color for the Light theme)

Test: manually verified SuspendedAppActivity theme on AndroidTV with
      and without change
      $ adb shell
      $ su
      # am start -n android/com.android.internal.app.SuspendedAppActivity
        --ei android.intent.extra.USER_ID 0 --es
	SuspendedAppActivity.extra.SUSPENDED_PACKAGE com.android.tv.settings
	--es SuspendedAppActivity.extra.DIALOG_MESSAGE "Test Message"
Bug: 112831365

Change-Id: I31b29e2b3ae7b0ff09e50f75e3288859c1799c4b

Former-commit-id: 1e5720eac4095027f810e73c9db07aa7e5c8188e