Added header to QS Edit view
The TileAdapter associated with the RecyclerView used to Edit
the Quick Settings was modified to contain a TextView as a Header.

Position arguments were corrected by 1 to support this. In particular
those of the Accessibility contexts.

Padding in the TextView as well as the RecyclerView was reduced to
minimize the movement of the tiles when switching into the Edit View.
This is not completely achieved and it is dependent on the size of
the font.

Change-Id: Ic89cec8cae15c238ac6e0b1c554c06c6b7086bab
Bug: 77326457
Bug: 36076720#24
Test: Manual - Screenshots

Former-commit-id: 39ec861009df657f9603e10efb3f0b8aacf84097