Remove suspected false positives from light greylist.
These APIs are only seen to be used by code in java packages
com.android.server.* which corresponds to AOSP code that has been linked
into applications.

Linking system server code into apps like this is unlikely to work, and
should not be expected to work. It's likely that this has been done
accidentally, and the code is not actually used at runtime.

Bug: 112826823
Test: m
Change-Id: I0a72458d66b6af515e1da68d449d9b166680c258

Former-commit-id: d29a6353f1e89fa6b96a9b2e7e27bad402686f2e