BatteryStats: Fix BleScan apportioning
The current BleScan apportioning formula has an error.
leftOverRxTimeMs is supposed to be constant at this point,
representing all the transfer time not already accounted
for by the scanning time. That figure is independent of which uids have
already been processed.
Note that updateWifiState has the same basic
formula (for Wifi accounting) but without this bug.

Bug: 113269219
Test: none
Change-Id: I600ac5e238580d951baa60ba02ed6d99ff48b6b1

Former-commit-id: f1d860dd8f26d2a72aaa3842c9b02e802cec4dbd