Add an enum file for stats event
The list of events in this file is meant to be used with Generic Atom
defined in atoms.proto in statsd. Instead of unique int, we would like
to log it with an enum that shows semantic meaning of events. Also, this
change re-sorts imports and removes unused imports and remove the
previous logging line to test GenericAtom.

Test: statsd_test
Change-Id: Ia9335e48ada68bee6d716ad4d7ac50238932da95

Former-commit-id: 0546d54fb44ab3bae0d6a4c0f2e3a0ebbdaebe1a