Experiment to be able to change the navigation bar thickness (1/2)
Use resource overlays to change the 2 dimension navigation_bar_height
and navigation_bar_frame_height to specify how large the navigation bar
is and how large its buttons are (also for width). There are two
overlays: first to slim the navigation bar to 75% and the other to have
floating navigation buttons.

go/navbar-experiments-doc

Test: manual
Bug: 113952590
Change-Id: I936c9fa49ba029b4174a22b7e35ee070a9e07e4b

Former-commit-id: 605ea474905e176449924d2879ad6e93672d532d