DO NOT MERGE: Reset launch start times when removing a process
When removing an exiting process that is running foreground activities, clear the launch times for
the current windowing mode. When an app process is removed, activities from the process may be
relaunched. In the case of forceStopPackageLocked the activities are finished before any window
is drawn, and the launch time is not cleared. This will be incorrectly used to calculate launch time
for the next launched activity launched in the same windowing mode.

Bug: 80084651
Test: adb shell am start -W com.amazon.mShop.android.shopping/com.amazon.mShop.home.HomeActivity  && sleep 1 && adb shell am force-stop com.amazon.mShop.android.shopping
Change-Id: I2c4f0716c922baa1ad209b4ea1fa7ce366e0e108
(cherry picked from commit 96ef4b8b7122176906c7109ee7c8bfb922fcf6f1 [formerly 9372683dd2b6ceafee4714770b51e67066e3d8bc])

Former-commit-id: fe0e09e4e384ec26a4a6f5ffba1f6e60dae2a7f4