Don't Use NPE for Flow Control in getCellLocation
Remove catch of NPE from getCellLocation and do explicit
null checks. This will make it easier to differentiate
between actual problems and local NPEs. In pre-P, there
were multiple conditions that would intentionally return
null over the binder interface. In Post-P null is returned
when location is disabled, which means that null is a very
common return value. We shouldn't use NPE to handle the
case where location is disabled.

Bug: 114294321
Test: manual w/ SL4A (no automated tests for TM)
Change-Id: Ie5b3afa56f280ff901ff9fa7c93f958d754f109b

Former-commit-id: edf6ccaeb3c66baa331aa55fe0fcdc5c4697f3e7