DO NOT MERGE: Use GL_LINEAR filter when drawing scaled TextureView
Use GL_LINEAR instead of GL_NEAREST sampling, when drawing
a TextureView, which has a buffer size that does not match
layer size.
Scale SkImage to layer size with a matrix, instead of
passing wrong size to MakeFromTexture.
This CL must not be merged in master, becase this issue has
been fixed already by ag/4936023.

Bug: 114324288
Test: Passed CtsViewTestCases and CtsUiRenderingTestCases
Test: Ran apps using scaled TextureView including Instagram
Change-Id: I9ee659d08998c932d8b708a64f3c879ab696fd9e

Former-commit-id: 93ac4012f2453e9a68a554ebba266f3ddd03d9f0