Extract getSuitableLocalesForSpellChecker()
When I needed to introduced a new helper method for
TextServicesManagerService (TSMS) to fix Bug 22042994 [2] I was just
lazy and decided to put it into InputMethodUtils instead of creating a
new utility class for TSMS.

Now that InputMethodUtils is going to be moved to
com.android.server.inputmethod as a package-private utility class, it
is time to pay off my laziness.

With this CL, getSuitableLocalesForSpellChecker() will be extracted
out from InputMethodUtils to a new package-private utility class
com.android.server.textservices.

This is purely a mechanical refactoring.  There should be no
observable behavior difference.

 [1]: I094f1c33430f7904a1dac6167431d6df64a07212
      fe239fc31e6e1078ddd33e2c3c4d788395dbf13f [formerly 174843afb629c57af19e14ee3ec4a91358061dd9]

Bug: 77730201
Bug: 114660660
Bug: 115516399
Test: atest FrameworksCoreTests:com.android.internal.inputmethod
Test: atest FrameworksServicesTests:com.android.server.textservices
Change-Id: If6cd051aa0ae7780631d5bec5b7d01faac4a83a0

Former-commit-id: 3781ca8c64bf5e912f3ecefcc6b423fe47c670d0