Move ShadeViewRefactor to right package to fix build breakage.
ShadeViewRefactor's package java.lang.annotation was inconsistent
with its source directory; further, it broke taimen-userdebug-jdk9
because java.lang.annotation is a core library package and it's
no longer supported for framework to have code in such packages.

This CL fixes compilation by moving the annotation to the correct
package which corresponds to its source directory.

Test: make SystemUI-core
Bug: 115604102

Change-Id: I600e57da316f60855d71f69b1aa1939f2081d513

Former-commit-id: 23957aa974f29f6186f07db274bd8864f1ed9719