Update the reference profile from .dm files only during installs
In order to speed up first boot time we only create the current profile
and do not update the content of the reference profile. A system image
should already be configured with the right profile keys and the
profiles for the speed-profile prebuilds should already be copied (that's
done in #performDexOptUpgrade).

Test: manual (install apps and splits, take OTA)
Bug: 114046578
Change-Id: I99db76684430b929c52634c25b457e874648e205

Former-commit-id: 5c9a0085a3366b4c7f00586dba9691fac9016f22