Fix test failure in NonPhoneDependencyTest.java.
The recent change to NotificationGroupManager to use
HeadsUpManager.isAlerting over the row causes a NPE to occasionally
surface if NonPhoneDependencyTest runs long enough.  When it sets up
the listener, it eventually gets a callback that adds a notification and
calls into NotificationGroupManager even though it hasn't had its
dependencies properly set up.  This CL fixes that by setting its
HeadsUpManager to the mock.

Test: ensure in all tests that NotificationGroupManager has its
dependencies set when NotificationEntryManager adds a notification
Fixes: 116495228

Change-Id: I7b82eedbf46fde5c8709dd03fdd36059041ef911

Former-commit-id: c87a66b1cc21a09b4a0f549df2a63ebcca3a2a0e