Fix user setting for dark mode not being applied
There was nothing making it so that night mode would be read
from settings global when the user was switched. Additionally
the settings was not being explicitly saved for the user. This
change makes it so we save the setting per user and registers
a receiver to update the UI when the user changes.

Test: Manual verification in settings by toggling
Change-Id: I4ddbc6bbcc51c81cb0271525a71b37440581feab
Fixes: 116001869

Former-commit-id: 2f792ef813db75cf697f57845896a24c433a1dec