startop: Add iorapd binder integration test
Connects to the binder 'iorapd' service as a client
and tests round-trip callback functionality.

Also adds mockito-inline to the test build, allowing us to mock
any class/methods, even if it's static or final.

Note: selinux is currently not supported, need to use
 $> adb shell setenforce 0

Prior to running any tests relying on binder connectivity.

Bug: 72170747
Test: atest libiorap-java-tests
Change-Id: Id5ba579845105e331d764838c357924810a64c6c

Former-commit-id: 292a9347f721c1b665266993b8dd2d346966da16