Includes uid in dropbox entries for errors.
The uid is critical for separating which user a crash occurred on.
Currently, if an app is installed on multiple users, there is no way to
determine which user was running an app that crashed since we only have
the package name. The process id is transient, while the uid is assigned
when the app is installed.

BUG: 116871038
Test: Tested manually that dropbox entry for a crash contains uid.
Change-Id: I5fcd0c1a276717150549eef7ec9ec23afc280ecd

Former-commit-id: 12edf7bf0208d582bafeaea2d906f3df27ba2aec