ExifInterface: Fix checking possible overflow
Currently, in order to prevent overflow when retrieving
thumbnail image, ExifInterface checks if the given
thumbnail offset + thumbnail length value is greater than
the length of the file. Instead of using available(), which
may require rewinding the file to the beginning, this
CL creates a getter method for the length of the file.

Test: Run cts
Change-Id: Ic6f9cd4d0f1a075493c393ce3eacc832a79a99ee

Former-commit-id: 42d96ce87ce846b2c752f616f94ad37086f73433