Add synchronization to UiAutomationManager.
The client change method needs to be called with a lock
held, and there's a one-off bug that looks like two threads
were shutting down UiAutomation simultaneously. Adding
a lock to prevent this from happening.

Bug: 111170405
Bug: 110845380
Test: A11y CTS and unit tests. I can't repro the issue, so
I can't write a test that specifically causes the problem
that this test makes pass.

Change-Id: Ia01603fcca5bcee70efd24e7af667d47d3057d61

Former-commit-id: dd577717f4b0fdfadfe556624cd2b83dac2700d5