Remove Hidden TimestampType From CellInfo
The timestamp type in CellInfo was added as an
experiment when CellInfo was first created. There
was optimism that highly precise timestamps would
be useful and available; however, in practice no
vendors are known to support precise timestamps
from the modem. Having a consistent timestap that
has a reference source available to apps is more
valuable than this flexibility. Thus, the reference
will always be nanotime since Android system boot,
and the timestamps will always be stamped by the
Android framework when the records are retrieved
from the modem.

Bug: 36971551
Test: atest FrameworksTelephonyTests

Change-Id: Ic5340b6c03ddc36aa00e7694c85b3af27128652e

Former-commit-id: eb8e831c956db804402e547c8a8e45d43bf6a0e7