AM: Fix activity launch metrics for already visible activities 1/2
For a windowing mode, when an activity is launched, AML starts
measuring the delay until it is notified that an activity has drawn
and a app transition has started. If the activity is behind a floating
or translucent app, then AML will not receive a windows drawn
notification and the measurements will continue.
When another activity is launched, the first activity is treated as a
trampoline activity and the metrics are logged against the new
activity. This can cause long delay timings for hot launches.

Handle this case by discarding the metrics if a launching activity
is already visible.

Test: atest CtsActivityManagerDeviceTestCases:ActivityMetricsLoggerTests
Bug: b/117148004
Change-Id: I6985099d3126db634d630897d2bced360f0c73d3

Former-commit-id: f8accc5b307e3ec2dc1b99acd9fd5358b9a26dd9