Makes bluetooth profile services bind to current (foreground) user.
If profile services are created from a singleton process
(like SystemUI), they need to be re-bind to a new foreground user
when we switch to a secondary user.
This is achieved by binding to UserHandle.CURRENT_OR_SELF.

If the process doesn't have adequate permissions to bind to current,
it will bind to self (which is the same as the previous behavior).

Change-Id: Ib1134a1a62887d6f5b2c97301dccd3223ade9ed2
Fixes: 117517812
Test: ran the existing core bluetooth tests; manual testing on a mojave, verifying that the service correctly rebinds from sys ui after user switch.

Former-commit-id: 2a9131f76879cbb7cc572ba999428197aece7e18