Removed references to ActivityRecord and TaskRecord in AM (24/n)
Once ActivityRecord and TaskRecord are moved to the WM package they
can no longer be accessed by classes in the AM package. This CL fixes
up some of the access before the move.
- Moved activity relaunch reasons to ActivityTaskManagerService and made
them public to allow access from the AM side in the future.
- Use task id to reference task in AppErrorDialog and AppErrors vs. the
TaskRecord object.
- Also, removed out-of-date code that tries to re-start a task if it
can't start in through recents. The reent start also looks for the task
in the stacks now so if the task exists it will be started through that
channel and we no longer need this fallback.

Test: Existing test pass
Bug: 80414790
Change-Id: I8df76a4f107fa526d726d7522eb5fcee6958b45d

Former-commit-id: 64258367f4afaeb37ec60f54406023fd19f27b04