Temporarily revert device identifier access check to previous behavior
If a calling package does not meet the new requirements for device identifier
access the calling package and method will be logged and the previous
READ_PHONE_STATE permission check will be performed to grant access to the
requested identifier. This is to prevent additional breakage for apps that
currently require device identifiers but have not yet been granted the
privileged permission or carrier privileges.

Bug: 117585389
Test: cts-tradefed run cts -m CtsPermissionTestCases \
      -t android.permission.cts.TelephonyManagerPermissionTest
Test: Manually invoked an app targeting pre-Q and verified access
      to device identifiers with the READ_PHONE_STATE permission.

Change-Id: I03339486a2d6971b93472479b79959c888beba1e

Former-commit-id: 9db22c7a89ca50c0a2e41df08baef56f4618efcd