Catch SecurityException in EuiccCardController and trigger callback.
Currently, if checkCallingPackage() throws a SecurityException, it won't
trigger the callback. Since, EuiccCardManager is async which will return
after invoke EuiccCardController, the throwed SecurityException will get
lost. In this CL, the exception is caught in EuiccCardController and
triggers the callback in this case.
Also, fix a typo in EuiccCardManager.

Bug: 78599148
Bug: 110845244
Test: test on phone
Merged-In: I6487d7b29bc9ec0d2ead482ac7d7804b5325a379
Change-Id: I6487d7b29bc9ec0d2ead482ac7d7804b5325a379

Former-commit-id: 35063d68fa92f3fc241c1209d556be7a9d9770ce