Enable @SoftInputModeFlags annotation in IMMS
This is a preparation to split
com.android.internal.view.InputMethodClient.

When WindowManager.LayoutParams.SoftInputModeFlags annotation was
originally introduced [1], somehow we could not use that annotation in
InputMethodManagerService (IMMS) probably because of some build system
or compiler limitation since the attribute belongs to frameworks.jar
while IMMS belongs to services.jar.

Now it seems that such a limitation no longer exists, there is no
reason to not enable @SoftInputModeFlags annotation in IMMS.

@SoftInputModeFlags is @Retention(SOURCE) anyway. Hence there should
be no behavior difference at run time.

 [1]: I96300b090edce327d0515c740183abe91ded6bac
      47937f8467681c5debfe55030a61bfc8d75cf7fa [formerly 22dac1c8df4ec212e8195a69d2de15d313d724fb]

Bug: 118040692
Test: compile
Change-Id: I5fb04a9c63d7123d25f2dbcc7e24b51ca79687ee

Former-commit-id: b0c26453771970978f31f9592123fb91cbf51a56