WM: Fix jumpcut animations
ROTATION_ANIMATION_JUMPCUT behaves like ROTATION_ANIMATION_SEAMLESS. In some cases we want to
simply freeze the screen while the app draws in the newrotation. This will revert back the
previous behavior.

Also fixes a flicker seen during jumpcut animation by fixing the inital position of the black
frame. The position was not set for Enter black frames when it was created.

Bug: 118169297
Test: atest FlickerTest:JumpcutAppRotationTest
Test: manually test with unity test app and verify jumpcut animation works
Change-Id: I87fd2b8536102e1c26adcf6ae16279aeb68d1e53

Former-commit-id: a1d1481f90e51bcaca4e0d2682c50645b5a0beeb