Extract IMM internal flags into StartInputFlags
This is a mechanical refactoring like we did to split
InputMethodClient for Bug 118040692.

So-called "controlFlags" in InputMethodManager (IMM) was originally
introduced for IMM#startInput()/windowGainedFocus() [1] to carry
additional client information then reused when we unify startInput()
and windowGainedFocus() into startInputOrWindowGainedFocus() [2].

This CL mechanically moves the location where those flags are defined,
from InputMethodManager.java to a newly created file
StartInputFlags.java.

This is still just a mechanical refacotring / renaming.  There should
be no user-visible behavior change.

 [1]: Icb58bef75ef4bf9979f3e2ba88cea20db2e2c3fb
      21f8a4e18deee8383a982c3f28395f5a46f0ee4b [formerly 7663d80f6b6fd6ca7a736c3802013a09c0abdeb9]
 [2]: I56934f18e30d90fcdf77bcbb0c35a92a5feb1b82
      1a41a3bc1bbacb40045ba7f0c4586f1ce93bdd87 [formerly 05c25f8a3a033816ac25aa5cd7db5b1ab495bc3f]

Test: atest CtsInputMethodTestCases CtsInputMethodServiceHostTestCases
Test: atest FrameworksCoreTests:com.android.internal.inputmethod.InputMethodDebugTest
Change-Id: If5a4810dece852edcff0d1119463711249bb7ef2

Former-commit-id: 35fa6d581eba553218946056e828e496324a5246