Fix port id mismatch temporarily.
ag/5409198

On Atom, HDMI1 port id is 2. HDMI2 is 4. HDMI3 is 1.
They are different from the port index.

When we do manual switch, we use port id to set Routing Port.
When we receive ActiveSource/SetStreamPath, we use port index.
There is a mismatch right now before we transit from hard code to a
portid-tvinputid mapping.

Note this is a temporary solution to avoid potential bug.

Test: manual.
Change-Id: I949587e17b815a755abf7cb17a70da1262efcd70

Former-commit-id: 0063811f183504f286e5e731cee85ab7069fd121