More robust persisting of screenshots.
Start using the newly added APIs for contributing new media.  This
gives us strong lifecycle around the "pending" state of the
screenshot while it's being written, before we eventually publish or
abandon it.

All the previous ImageColumns fields are now filled in
automatically by MediaProvider once the image is persisted.

Bug: 115377970, 118167155
Test: atest android.provider.cts.MediaStorePendingTest
Change-Id: I0bd05336f3a309007c6d0f4fb711c5b467bc18ea

Former-commit-id: 7bba5072d48a5e6dcce189ca900aeafd980b7ef4