Clarifying some battery saver policy variables.
The flags for the network firewall, screen brightness adjustment, and
data saver are essentially to decide whether to turn on these features
or not, so having them be *_disabled is confusing for someone to read.
These flags differ from others such as animation_disabled or
vibration_disabled, which are actually turning off features instead of
enabling them. I'm updating the internal variable names and adding
documentation to make the behavior clearer. I haven't changed the flag
names themselves since that could make things a little more difficult if
someone wants to set flags later on.

Bug: 117126975
Test: atest com.android.server.power.BatterySaverPolicyTest
Change-Id: If3fa220f9f62117eac39b54bf771a9cc17bcc911

Former-commit-id: 7e5c7018d335054be574eb351f92d94beb82d6bf