Add a Notifications API for contextual (smart) actions.
With this CL we add a new semantic action to Notification.Action to
represent contextual actions - i.e. actions specific to the current
notification. Contextual actions will be displayed differnently to
regular actions.

If the app itself does not provide smart replies or smart actions the
framework will generate smart replies and actions and put those in the
notification.

The UI for smart actions will be added in a separate CL.

With this CL we also remove a couple of unnecessary fields, namely:
NotificationInflater.mSmartActions
NotificationInflater.AsyncInflationTask.mSmartActions
the actions stored in these fields are stored in
NotificationData.Entry anyway.

Bug: 119010281
Test: write an app adding an Action with the new semantic type and
ensure that (contextual) action doesn't show up with the other actions.
Test: atest SystemUITests

Change-Id: Icd192fe2273b0474729dd4e413e0a6c2d75ff0fa

Former-commit-id: 1463d83d4e5e3096dd1d56f854df46dae02c35cb