Put Vulkan WebViews on a HW layer if stencil clip
Put WebViews in a HW layer, if the clip is a non-rect.
This CL reuses logic implemented by ag/705975. This used
to be the way GL WebViews were drawn in Android M.
Implement complex clip detection at recording
time, which was not previously supported by the canvas.
Vulkan WebViews using GL interop are already drawn
in a layer, but this CL will be useful when WebView
supports new Vulkan interop.

Test: WebView CTS pass for Vulkan and GL
Bug: 115613038
Change-Id: I9b02c6f4de8efd504a7507633f3d849004215a16

Former-commit-id: f09ee58eb7edf7efb0cf8e24ad763a582dac0228