Added turn on/off virtual display
VirtualDisplays are considered on if they have a surface and off if they don't.
This causes problems for displays that don't have a surface. Instead,
create a separate API to allow the client to request the surface be on
or off.

To ensure backwards compatibility for VirtualDisplays, set the display
to on or off when the surface is added or removed.

Change-Id: If9d8db94a66d6484ac492a53c1cd8fb7da851b88
Fixes: 119209373
Test: Created surfaceless virtual display that can be manually turned on or off

Former-commit-id: da4c694bffbf14910b34df8370f89751c1aa7ea2