Hide the recents button on the secondary display
The recent only support on the default display now. If users
launch the recent on the secondary display,the recent will
show on the default display.Hide the recents button of the
navigation bar on the secondary display to prevent the weird
behavior first.

Bug: 117749252
Test: atest SystemUITests
Test: atest NavigationBarButtonTest
Test: manual - Create a virtual display with system decoration
and check if the recents button on the navigation bar.
Change-Id: I6f6111d0043046c192cdc6ad2b1fd7d075cb9d26

Former-commit-id: 2aaa91aa69f2fcdc23c3226c3ddc87e4432af681