WifiConfiguration: Add a field to mark network as trusted
The |ephemeral| field (added for WFA) is currently used for 2 purposes:
a) Network is not persisted & network config
is automatically removed at the end of the connection.
b) Network is marked untrusted (in NetworkInfo capabilities) after connection.

For the new network request API, we want (a), but not (b) (Network should
be marked trusted because user explicitly allowed it).

Add a new flag to explicitly mark the network as trusted or not. Hence
(a) & (b) are now indicated by different flags.

Bug: 113878056
Test: ./frameworks/base/wifi/tests/runtests.sh
Change-Id: If4da84ca4fc03fd6a835b71a3e18e966c3b11f05

Former-commit-id: db99b31cfa48ebfdaa711bd0c18d589a6b7cede1