[DO NOT MERGE] AM: Fix invalid hot launch app metrics seen with quickstep
When quickstep is enabled, and the home button is pressed, AML receives two launches. The second
launch is always incomplete. When an activity is then launched, the Launcher activity is
treated as a trampoline activity and the launch will be reported as a hot launch with inflated
times.

The second launch is incomplete because we never get a window's drawn message for the second
launch. We solve this by keeping track of windows drawn state in the Activity Record. If an
activity is launched and its windows are already drawn, then we abort the launch metrics.

Test: manual test with quickstep
Test: atest ActivityMetricsLoggerTest
Bug: 119221454
Change-Id: Id34a1c42c9b2ea356757109fedf5e2a1a3b020c8

Former-commit-id: c1d2ff60f0f95cdfdfe8d89c9326ae5f16e7adc4