Bypass most session work for APEX packages
Fixes a merge error introduced in the multi-package API refactor that
causes all APEX installs to fail as they were being treated as full
APKs while making the session active for handling by
PackageManagerService. This change bypasses that work if the package
is being installed as APEX as the package will not be passed to
PackageManagerService.

Change-Id: Ic7534e49e4a552b64cc1e046ce79272ece1be73a
Fixes: 119411166
Test: adb install apex.test.apex

Former-commit-id: a77a677c385319edfd32abd8017f9f10f52594b6