Sending a TIME_TICK on screen_on
We used to send a single TIME_TICK on coming out of idle but now that is
delayed for a while after screen_on, which can lead to obsolete ui for
components relying on TIME_TICK for updating themselves.
Also, added a tag for TIME_TICK listener so it's easier to track in the
dump.

Test: atest FrameworksMockingServicesTests:AlarmManagerServiceTest

Bug: 119328963
Change-Id: I7740236c5c6dfd3c5338856c8e939a9c42bd519e

Former-commit-id: 12bd016ded0fb2750b9de4665b61ed2b8230ac1a