Let low-ram devices override voice recognisers on
The top-level framework overlay already says this should (by default) be
skipped if the device is low-ram (leaving the device able to turn it
back on again without committing to a specific voice recognizer at build
time):
  <feature name="android.software.voice_recognizers" notLowRam="true" />

The form factor-specific overlay for tablets does override this back to
a state where the feature is always enabled regardless of low-ram:
  <feature name="android.software.voice_recognizers" />

But this is probably a mistake as the code before this change ignored
the feature flag anyway in case of a low-ram device, using a hardcoded
check.

Lets Android TVs have voice recognisers working on all devices by
default without needing to set katniss' package as a force override.

Bug: 117630721
Test: flashall ; adb logcat | grep VoiceInteractionService
Change-Id: I7816bebbc363ae0751b097fe1b6cdc2a646b20e0

Former-commit-id: ba3af56aaae2d920d4e1593b0c71f7f55fd47061