Update comments for security type in AccessPoint
There are no resources in array.xml that references the security or
PSK field, so updating the comments to match.
There is also no need to keep the list sorted -- new values should be
added to the end and any string-array resource values can be added to
the end accordingly.

Test: None
Change-Id: I85e7de9c9f08468f6aaeaa9af1482e304a501a76

Former-commit-id: 7cdbe199fd0a506f0568b6000f316f3c783cf215