WifiScanner: Pass the band scanned in ScanData
Replace the existing |isAllChannelsScanned| variable with the exact band
that was scanned.
|isAllChannelsScanned| flag has been overloaded over the years to convey
both {2g + 5g} & {2g + 5g + dfs} scans. This was needed for wifi wake
and other clients for whom both were equivalent. For sending out
scan broadcasts to external apps, we need to identify scans with
{2g + 5g + dfs} only.

Also, add annotation for the permission required for |registerScanListener|.
This should only be available to wifi stack.

Bug: 119316884
Test: ./frameworks/base/wifi/tests/runtests.sh
Test: Will send for integration tests
Change-Id: Iea50851ef9f1e98d2e06601ba89d112cd2dd4e38

Former-commit-id: 5a600f5d40073aa35112e0b6835e37b821d7ab39