Remove grouping for platform permissions
The grouping was never meant to be authorative, but this was not
documented. This lead to people assuming that it would not change and
implemented code on top of it.

Further handling permissions and groups requires the app to handle all
kind of special cases which is not obvious. By not exposing the grouping
the app developer will become aware that she/he should not handle
permissions.

Test: atest CtsAppSecurityHostTestCases:android.appsecurity.cts.PermissionsHostTest
Change-Id: Iad1480efde338fa8c1cfba6372b1552c5d2361c0

Former-commit-id: 17eae45cf9a3948ed268e51bf13528ad82a465f0